requirements.txt
mne>=1.6.0
numpy>=1.24.0
scipy>=1.11.0
pandas>=2.0.0
scikit-learn>=1.3.0
torch>=2.1.0
wfdb>=4.1.2
tqdm>=4.66.0
pyyaml>=6.0.1
matplotlib>=3.8.0

pyserial>=3.5
bleak>=0.22.0
